requirements.txt
langchain-astradb
langchain
langchain-openai
pypdf
python-dotenv
datasets
flask
google-generativeai
langchain-google-genai
langchain-community
-e .